+++ noodles-bam/src/reader/record.rs
@@ -13,7 +13,7 @@ use bytes::Buf;
 use noodles_sam as sam;
 
 use crate::{
-    record::{Cigar, Data, ReferenceSequenceId},
+    record::{Data, ReferenceSequenceId},
     Record,
 };
 
@@ -179,21 +179,45 @@ where
     Ok(())
 }
 
-fn read_cigar<B>(buf: &mut B, cigar: &mut Cigar, n_cigar_op: usize) -> io::Result<()>
+fn read_cigar<B>(buf: &mut B, cigar: &mut sam::record::Cigar, n_cigar_op: usize) -> io::Result<()>
 where
     B: Buf,
 {
+    use sam::record::cigar::{op::Kind, Op};
+
+    fn decode_cigar_op(n: u32) -> io::Result<Op> {
+        let kind = match n & 0x0f {
+            0 => Kind::Match,
+            1 => Kind::Insertion,
+            2 => Kind::Deletion,
+            3 => Kind::Skip,
+            4 => Kind::SoftClip,
+            5 => Kind::HardClip,
+            6 => Kind::Pad,
+            7 => Kind::SequenceMatch,
+            8 => Kind::SequenceMismatch,
+            _ => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidData,
+                    "invalid CIGAR op kind",
+                ))
+            }
+        };
+
+        let len = n >> 4;
+
+        Ok(Op::new(kind, len))
+    }
+
     if buf.remaining() < mem::size_of::<u32>() * n_cigar_op {
         return Err(io::Error::from(io::ErrorKind::InvalidData));
     }
 
-    let cigar = cigar.as_mut();
-
-    cigar.resize(n_cigar_op, Default::default());
     cigar.clear();
 
     for _ in 0..n_cigar_op {
-        cigar.push(buf.get_u32_le());
+        let op = decode_cigar_op(buf.get_u32_le())?;
+        cigar.as_mut().push(op);
     }
 
     Ok(())